analyzer: clear remaining real-concern bugprone-* clusters in core
Step 7 from docs/analyze-triage.md. Filters the non-step-6 source
files for actionable bugprone-* categories and clears every
finding except a 67-hit reserved-identifier residue in
gtk_hlist_compat.c (all GLib macro-expansion noise — \_\_inst /
\_\_r / \_\_t in g_return_if_fail and GTK_IS_* invocations).

By category:

- bugprone-assert-side-effect (chat.c, login_packet.c): NOLINTs
  on g_assert checks where clang-tidy mis-reads a -> chain or
  pointer compare as a side effect. Sizeof on a fixed array
  doesn't evaluate its operand, and a < / == against a local
  is pure.
- bugprone-integer-division (gtk_hlist_compat.c): hlist row text
  y-centring (cell_area->height - th) / 2 done in float so
  subpixel positions survive to the snapshot.
- bugprone-not-null-terminated-result (login_packet.c x2):
  NOLINTs on the two memcpys that write length-prefixed binary
  wire fields (name list entry, 4-byte AppID).
- bugprone-branch-clone (tasks.c): merged -127 (tracker cancel)
  and -129 (tracker quit) branches with ||. xfers.c: removed the
  dead-pre-stats nr_gets/nr_puts hook left over from the original
  XFER_GET/XFER_PUT dispatch, which had two identical-looking
  empty branches.
- bugprone-redundant-branch-condition (rcv.c): dropped the inner
  if(uid && uid == htlc->uid) that restated the outer.
- bugprone-misplaced-widening-cast (xfers.c): widen fi.comlen to
  ssize_t before adding the 133 constant so the comparison
  against htxf_io_write's return value is in the wider type.
- bugprone-suspicious-string-compare (files.c x2, options.c):
  explicit " != 0" on memcmp / strcmp results that were used as
  truthy values.
- bugprone-signed-char-misuse (connect.c, network.c, rcv.c,
  usermod.c): connect.c::len_addr re-typed as unsigned char so
  the length byte we read off the wire isn't sign-extended into
  size_t. network.c::ctx->secure assignment casts through
  unsigned char (the param itself is still 'char' to keep the
  caller signatures stable; that lives in connect.c). rcv.c dir
  walk casts *p through guint8 before the dir_char compare.
  usermod.c bitno write casts through unsigned char.
- bugprone-reserved-identifier (commands.c, usermod.c, xfers.c):
  \_\_commands → commands_tbl in commands.c; \_\_uesp → uesp
  in usermod.c; \_\_arg → arg across five xfers.c thread-fn
  parameters (and xfers.h's prototype).

Build clean; 50 unit + proto tests green.

+++ src/gtk_hlist_compat.c
@@ -536,7 +536,9 @@ gtk_hlist_overlay_cell_snapshot (GtkCellRenderer *cell, GtkSnapshot *snapshot,
         }
 
         tx = (float)(cell_area->x + x_off);
-        ty = (float)(cell_area->y + (cell_area->height - th) / 2);
+        /* Vertical-center the text in the cell; keep the divide in
+         * floating point so subpixel positions survive to the snapshot. */
+        ty = (float)cell_area->y + (float)(cell_area->height - th) / 2.0f;
 
         /* Optional 4-direction outline. Draws the same layout four
 		 * times at ±1 px offsets in pure black before the fg draw
